requirements.txt
accelerate==1.10.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
filelock==3.20.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
psutil==7.1.0
pvporcupine==3.0.5
PyAudio==0.2.14
pycparser==2.23
pyserial==3.5
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
schedule==1.2.2
sentencepiece==0.2.1
setuptools==80.9.0
sounddevice==0.5.2
srt==3.5.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
